Move the class to base_writer.py
